+++ app/ide-desktop/lib/dashboard/src/layouts/Editor.tsx
@@ -57,20 +57,7 @@ export default function Editor(props: EditorProps) {
     if (hidden) {
       return
     } else {
-      const params = {
-        platform: detect.platform(),
-        architecture: detect.architecture(),
-      }
-      const gtagEventCurrent = gtagEventRef.current
-      gtagEventCurrent('open_workflow', params)
-      const onBeforeUnload = () => {
-        gtagEventCurrent('close_workflow', params)
-      }
-      window.addEventListener('beforeunload', onBeforeUnload)
-      return () => {
-        window.removeEventListener('beforeunload', onBeforeUnload)
-        gtagEventCurrent('close_workflow', params)
-      }
+      return gtagHooks.gtagOpenCloseCallback(gtagEventRef, 'open_workflow', 'close_workflow')
     }
   }, [projectStartupInfo, hidden])
